Create utility class to get user's forum title
